raise puzzle when alarm starts (unless in LED clock mode); store and restore original ALSA state
The puzzle is either shown in the existing ffalarms window which is
raised (this unfortunately does not work on illume home screen) or a
new ffalarms with the puzzle is started.

Previously the stereoout.state was restored after the alarm so if
alarm interrupted a phone call, then the phone call was made silent on
both ends even after stopping the alarm. Now there is a potential
reverse problem: if the phone call ends before you stop the alarm then
the alarm will restore now useless phone call suitable ALSA state
saved when the alarm started.


git-svn-id: svn+ssh://svn.projects.openmoko.org/svnroot/ffalarms/trunk@2 4d22690b-c26c-4996-9a96-d273cef9f366